requirements.txt
fastapi
uvicorn
python-multipart
fpdf2
Pillow
pytz
requests